agent/controller/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "controller"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlmodel>=0.0.19",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "grpcio>=1.64.0",
    "grpcio-tools>=1.64.0",
    "grpcio-reflection>=1.64.0",
    "greenlet>=3.0.0",
    "cryptography>=42.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
